+++ js/chat_app/src/App.js
@@ -1,33 +1,35 @@
 import Login from "./components/Login";
 import DashBoard from "./components/DashBoard";
-import { useState } from "react";
-import {
-  BrowserRouter as Router,
-  Navigate,
-  Route,
-  Routes,
-} from "react-router-dom";
+import { useState, useEffect } from "react";
+import { Navigate, Route, Routes, useNavigate } from "react-router-dom";
 import { UserContext } from "./components/UserContext";
 
 function App() {
   const session = JSON.parse(sessionStorage.getItem("user-info"));
   const [auth, setAuth] = useState(session);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const auth = JSON.parse(sessionStorage.getItem("user-info"));
+    setAuth(auth);
+    if (auth !== null) {
+      navigate("/dashboard", { replace: true });
+    }
+  }, []);
 
   return (
     <div>
       <UserContext.Provider value={{ auth, setAuth }}>
-        <Router>
-          <Routes>
-            <Route
-              path="/"
-              element={
-                <Navigate to={auth ? "/dashboard" : "login"} replace={true} />
-              }
-            />
-            <Route path="/login" element={<Login />} />
-            <Route path="/dashboard" element={<DashBoard />} />
-          </Routes>
-        </Router>
+        <Routes>
+          <Route
+            path="/"
+            element={
+              <Navigate to={auth ? "/dashboard" : "/login"} replace={true} />
+            }
+          />
+          <Route path="/login" element={<Login />} />
+          <Route path="/dashboard" element={<DashBoard />} />
+        </Routes>
       </UserContext.Provider>
     </div>
   );
